+++ cmd/gen_docs.go
@@ -30,6 +30,7 @@ func genDocsCmd(longAppDesc string) *cobra.Command {
 				editCmd(&config, &configErr),
 				listCmd(&config, &configErr),
 				describeCmd(&config, &configErr),
+				checkCmd(&config, &configErr),
 				genCmd(),
 			)
 			core.CheckIfError(err)